Introduced a FileSystemWatcher and replaced ListView with table in SessionListView.
When a new data file is added or an existing data file is deleted, the stats will be updated automatically. For this, the way data is loaded and load progress is shown had to be refactored.
As the ListView used by SessionListView did not give enough possibilities on how the session info can be displayed, it was replaced with a table.
Additionally, a few bugs were fixed and a workaround for missing MVP information in the data file was added.